MT8362: auth_finder.pl: improve multiple subfields management
Before this patch, choosing an authority which have multiple subfields
with the same tag, values were joined in a single subfield, seperating
values with a pipe ('|').
Now, if an authority has for example two $x subfields, there will be two
$x subfields in the biblio too. The exception is when the subfield is
not repeatable in the biblio marc framework. In this case, only the
first subfield is copied.